// src/types/passenger.ts
export interface Passenger {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  nationality: string;
  passportNumber: string;
  passportExpiry: string;
  emergencyContactName: string;
  emergencyContactPhone: string;
  specialRequirements: string;
  spaceflightConsent: boolean;
} 